+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 * Reworked M.A. Jump
   * Now called Double Jump, it's one level only and is a treasure enchantment
+* Moved Critical enchantment to Reborn
 
 ## 1.9.14
 * Soul blast furnaces now take 50% more time to smelt stuff (compared to blast furnaces) instead of 100%